remove extra tests in http module
