+++ test/test_connection.py
@@ -4,6 +4,7 @@ import stardog.admin
 import stardog.connection as connection
 import stardog.content as content
 import stardog.content_types as content_types
+import stardog.exceptions as exceptions
 
 
 @pytest.fixture(scope="module")
@@ -27,12 +28,12 @@ def admin():
 
 
 def test_transactions(conn, admin):
-    data = '<urn:subj> <urn:pred> <urn:obj> .'
+    data = content.Raw('<urn:subj> <urn:pred> <urn:obj> .', content_types.TURTLE)
 
     # add
     conn.begin()
     conn.clear()
-    conn.add(content.Raw(data, content_types.TURTLE))
+    conn.add(data)
     conn.commit()
 
     assert conn.size() == 1
@@ -46,7 +47,7 @@ def test_transactions(conn, admin):
 
     # rollback
     conn.begin()
-    conn.add(content.Raw(data, content_types.TURTLE))
+    conn.add(data)
     conn.rollback()
 
     assert conn.size() == 0
@@ -70,6 +71,27 @@ def test_transactions(conn, admin):
     conn.commit()
 
     assert conn.size() == 0
+
+    # add named graph
+    conn.begin()
+    conn.add(data, graph_uri='urn:graph')
+    conn.commit()
+
+    assert conn.size(exact=True) == 1
+
+    # remove from default graph
+    conn.begin()
+    conn.remove(data)
+    conn.commit()
+
+    assert conn.size(exact=True) == 1
+
+    # remove from named graph
+    conn.begin()
+    conn.remove(data, graph_uri='urn:graph')
+    conn.commit()
+
+    assert conn.size(exact=True) == 0
 
 
 def test_queries(conn, admin):
@@ -158,6 +180,48 @@ def test_queries(conn, admin):
     assert len(q['results']['bindings']) == 0
 
     conn.commit()
+
+
+def test_reasoning(conn, admin):
+    data = content.Raw(
+        b'<urn:subj> <urn:pred> <urn:obj> , <urn:obj2> .',
+        content_types.TURTLE
+    )
+
+    # add
+    conn.begin()
+    conn.add(data)
+    conn.commit()
+
+    # consistency
+    assert conn.is_consistent()
+
+    # explain inference
+    r = conn.explain_inference(
+        content.Raw('<urn:subj> <urn:pred> <urn:obj> .', content_types.TURTLE)
+    )
+    assert ('status', 'ASSERTED') in r[0].items()
+
+    # explain inference in transaction
+    conn.begin()
+    conn.add(content.Raw('<urn:subj> <urn:pred> <urn:obj3> .', content_types.TURTLE))
+
+    with pytest.raises(exceptions.StardogException):
+        r = conn.explain_inference(
+            content.Raw('<urn:subj> <urn:pred> <urn:obj3> .', content_types.TURTLE)
+        )
+        assert len(r) == 0
+
+    # explain inconsistency in transaction
+    with pytest.raises(exceptions.StardogException):
+        r = conn.explain_inconsistency()
+        assert len(r) == 0
+
+    conn.rollback()
+
+    # explain inconsistency
+    r = conn.explain_inconsistency()
+    assert len(r) == 0
 
 
 def test_docs(conn, admin):
